refactor: move App component and styles to src/components directory
- Rename 'src/App.tsx' to 'src/components/App.tsx'
- Rename 'src/App.styles.ts' to 'src/components/App.styles.ts'
- Update import paths accordingly
